backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.9
email-validator==2.1.1
google-auth==2.29.0
requests==2.31.0
razorpay==1.3.0
